+++ docker-compose.yaml
@@ -1,57 +1,3 @@
-# version: '3'
-
-# services:
-#   # ui:
-#   #   build:
-#   #     context: ./
-#   #     dockerfile: ./docker/client.Dockerfile.dev
-#   #   image: qaproject-ui:development
-#   #   platform: linux/x86_64
-#   #   volumes:
-#   #     - ./ui:/app/ui:delegated
-#   #     - /app/ui/node_modules/
-#   #   ports:
-#   #     - 8085:8085
-#   #   environment:
-#   #     - PORT=8085
-#   #     - HOST=0.0.0.0
-#   #     - NODE_ENV=development
-#   #   command: npm run --prefix ui start
-
-#   # server:
-#   #   build:
-#   #     context: ./
-#   #     dockerfile: ./docker/server.Dockerfile.dev
-#   #   image: qaproject-server:development
-#   #   platform: linux/x86_64
-#   #   depends_on:
-#   #     - mssql
-#   #   volumes:
-#   #     - ./server:/app/server:delegated
-#   #     - /app/server/node_modules/
-#   #     - ./.env:/app/.env
-#   #   ports:
-#   #     - 8086:8086
-
-#   mssql:
-#     # build: .
-#     image: mcr.microsoft.com/mssql/server:2019-latest
-#     ports:
-#       - 1433:1433
-#     volumes:
-#       - ./db_backup:/var/opt/mssql/backup
-#       - mssql-backup:/var/opt/mssql/backup
-#       - ~/GitHub/Data/mssql:/var/opt/mssql/data
-#       - ./restore-db.sh:/var/opt/mssql/scripts/restore-db.sh
-
-#     command: /bin/bash /var/opt/mssql/scripts/restore-db.sh
-#     environment:
-#       - ACCEPT_EULA=Y
-#       - SA_PASSWORD=mssql1Ipw
-
-# volumes:
-#   mssql-backup:
-
 # SQL COMMAND TO REBUILD DB
 version: '3'
