--- web/app/(feed)/featuree/[id]/page.tsx
@@ -1,8 +1,0 @@
-'use client'
-// import { useParams} from
-
-export default function page() {
-  // const { id } = useParams()
-
-  return <div>Modal</div>
-}
