--- web/app/features/[id]/edit/loading.tsx
@@ -1,5 +1,0 @@
-// 'use client'
-
-export default function loading() {
-  return <div>loading</div>
-}
