+++ web/app/layout.tsx
@@ -1,4 +1,3 @@
-import Link from 'next/link'
 import './globals.css'
 import AuthProvider from './AuthProvider'
 import NavMenu from './NavMenu'
@@ -23,19 +22,7 @@ export default function RootLayout({
         //  className={inter.className}
         >
           <NavMenu />
-
-          {/* <nav>Navbar in root layout</nav> */}
-          <div>
-            {/* <div className="w-full fixed top-0 flex flex-row items-center justify-between p-4">
-              <Link href="/todos">todos</Link>
-              <Link href="/features">features</Link>
-              <Link href="/testCases">testCases</Link>
-              <AuthCheck>
-                <SignOutButton />
-              </AuthCheck>
-            </div> */}
-            {children}{' '}
-          </div>
+          <div>{children} </div>
         </body>
       </html>
     </AuthProvider>
