+++ web/app/page.tsx
@@ -1,6 +1,3 @@
-// import Image from 'next/image'
-// import Link from 'next/link'
-
 export default function Home() {
   return (
     <div className="flex min-h-screen flex-col items-center justify-between p-24">
