--- web/app/testCases/TestCasesList.tsx
@@ -1,20 +1,0 @@
-import { TestCase } from '../../generated-api/models/TestCase'
-import TestCaseComponent from './TestCase'
-
-type TestCaseListProps = {
-  testCases: TestCase[]
-}
-
-export default function TestCaseList({ testCases }: TestCaseListProps) {
-  return (
-    <div>
-      {testCases.map((testCase) => {
-        return (
-          <div key={testCase.id}>
-            <TestCaseComponent testCase={testCase} />
-          </div>
-        )
-      })}
-    </div>
-  )
-}
