--- web/app/testCases/[id]/error.tsx
@@ -1,6 +1,0 @@
-'use client'
-
-export default function error(e: any) {
-  return <div>error: {JSON.stringify(e)}</div>
-  // return <div>error:</div>
-}
