--- web/app/testCases/error.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-export default function error(e: any) {
-  return <div>error{JSON.stringify(e)}</div>
-}
