--- web/app/testCases/loading.tsx
@@ -1,6 +1,0 @@
-// 'use client'
-// 'use-server'
-
-export default function loading() {
-  return <div>loading</div>
-}
