--- web/app/testCases/page.tsx
@@ -1,24 +1,0 @@
-import { TestCasesApi } from '../../generated-api/apis/TestCasesApi'
-import CreateTestCase from './CreateTestCase'
-import TestCasesList from './TestCasesList'
-
-const testCasesClient = new TestCasesApi()
-
-async function getTestCases() {
-  const response = await testCasesClient.testCasesGet({
-    ...{ cache: 'no-store' },
-  })
-  return response
-}
-
-export default async function TestCases() {
-  const testCases = await getTestCases()
-
-  return (
-    <div className="pt-32">
-      TestCases Page
-      <TestCasesList testCases={testCases} />
-      {/* <CreateTestCase featureId={2} /> */}
-    </div>
-  )
-}
