--- web/app/todos/[id]/error.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-export default function error() {
-  return <div>error</div>
-}
