--- web/app/todos/error.tsx
@@ -1,6 +1,0 @@
-'use client'
-// 'use-server'
-
-export default function error() {
-  return <div>error</div>
-}
